refactor: Reorganize project structure using feature-sliced architecture
