+++ GenHub/GenHub.Windows/Program.cs
@@ -5,8 +5,9 @@ using System.Runtime.InteropServices;
 using System.Threading;
 using Avalonia;
 using GenHub.Core;
-using GenHub.Services;
-using GenHub.ViewModels;
+using GenHub.Core.Interfaces.GameVersions;
+using GenHub.Features.GameVersions.Services;
+using GenHub.Common.ViewModels;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace GenHub.Windows;
