+++ GenHub/GenHub/App.axaml
@@ -1,13 +1,12 @@
 <Application xmlns="https://github.com/avaloniaui"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:GenHub"
+             xmlns:ui="clr-namespace:GenHub.Infrastructure.UI"
              xmlns:material="clr-namespace:Material.Icons.Avalonia;assembly=Material.Icons.Avalonia"
              x:Class="GenHub.App"
              RequestedThemeVariant="Default">
-             <!-- "Default" ThemeVariant follows system theme variant. "Dark" or "Light" are other available options. -->
-
-     <Application.DataTemplates>
-         <local:ViewLocator />
+             <!-- "Default" ThemeVariant follows system theme variant. "Dark" or "Light" are other available options. -->     <Application.DataTemplates>
+         <ui:ViewLocator />
      </Application.DataTemplates>
 
     <Application.Styles>
