+++ GenHub/GenHub/App.axaml.cs
@@ -2,17 +2,16 @@ using System;
 using Avalonia;
 using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Markup.Xaml;
-using GenHub.Core;
-using GenHub.Services;
-using GenHub.ViewModels;
-using GenHub.Views;
+using GenHub.Common.ViewModels;
+using GenHub.Common.Views;
+using GenHub.Core.Helpers;
+using GenHub.Features.GameVersions.Services;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace GenHub;
 
 public partial class App : Application
 {
-
     public override void Initialize()
     {
         AvaloniaXamlLoader.Load(this);
