+++ GenHub/GenHub/Common/Views/MainView.axaml
@@ -3,9 +3,9 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:material="clr-namespace:Material.Icons.Avalonia;assembly=Material.Icons.Avalonia"
-             xmlns:vm="clr-namespace:GenHub.ViewModels"
+             xmlns:vm="clr-namespace:GenHub.Common.ViewModels"
              mc:Ignorable="d" d:DesignWidth="1280" d:DesignHeight="720"
-             x:Class="GenHub.Views.MainView"
+             x:Class="GenHub.Common.Views.MainView"
              x:DataType="vm:MainViewModel">
 
     <Design.DataContext>
