+++ GenHub/GenHub/Common/Views/MainWindow.axaml
@@ -2,9 +2,9 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:GenHub.Views"
+        xmlns:local="clr-namespace:GenHub.Common.Views"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="GenHub.Views.MainWindow"
+        x:Class="GenHub.Common.Views.MainWindow"
         Title="GenHub"
         WindowStartupLocation="CenterScreen"
         ExtendClientAreaToDecorationsHint="True">
