+++ GenHub/GenHub/Infrastructure/UI/ViewLocator.cs
@@ -1,9 +1,9 @@
-﻿using System;
+using System;
 using Avalonia.Controls;
 using Avalonia.Controls.Templates;
-using GenHub.ViewModels;
+using GenHub.Common.ViewModels;
 
-namespace GenHub;
+namespace GenHub.Infrastructure.UI;
 
 public class ViewLocator : IDataTemplate
 {
